requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
Pillow==10.1.0
requests==2.31.0
numpy==1.24.3
